move remaining cleaning tasks to cleaner class
