+++ ewah/cleaner.py
@@ -7,6 +7,84 @@ from typing import List, Dict, Optional, Any, Callable, Union
 from copy import deepcopy
 from hashlib import sha256
 
+# Refactor me
+from bson.json_util import dumps # dumping mongob objects to string
+from collections import OrderedDict
+from decimal import Decimal
+
+import json
+
+class EWAHJSONEncoder(json.JSONEncoder):
+    """Extension of the native json encoder to deal with additional datatypes and
+    issues relating to (+/-) Inf and NaN numbers.
+    """
+
+    def default(self, obj):
+        """Method is called if an object cannot be serialized.
+
+        Ought to return a serializeable value for the object. Ought to raise an error
+        if unable to do so.
+
+        Implemented types:
+            - Decimal -> float
+            - bson.objectid.ObjectId -> string
+        """
+
+        if isinstance(obj, Decimal):
+            return float(obj)
+        # Let the base class default method raise the TypeError
+        return super().default(obj)
+
+    def iterencode(self, o, _one_shot=False):
+        """Overwrite the iterencode method because this is where the float
+        special cases are handled in the floatstr() function. Copy-pasted
+        original code and then adapted it to change floatstr() behavior.
+
+        This was necessary because the json module accepts (+/-) Infinity and NaN
+        objects as floats, but PostgreSQL uses the tighter JSON standard and does not
+        accept them in json. The json module offers no optional flag to deal with this
+        issue natively. Thus, overwrite the iterencode method and remove Inf and NaN
+        by returning null instead.
+        """
+        if self.check_circular:
+            markers = {}
+        else:
+            markers = None
+        if self.ensure_ascii:
+            _encoder = json.encoder.encode_basestring_ascii
+        else:
+            _encoder = json.encoder.encode_basestring
+
+        def floatstr(
+            o,
+            allow_nan=self.allow_nan,
+            _repr=float.__repr__,
+            _inf=float("inf"),
+            _neginf=-float("inf"),
+        ):
+            # Check for specials.  Note that this type of test is processor
+            # and/or platform-specific, so do tests which don't depend on the
+            # internals.
+            if not (o != o or o == _inf or o == _neginf):
+                return _repr(o)
+            if not allow_nan:
+                raise ValueError(
+                    "Out of range float values are not JSON compliant: " + repr(o)
+                )
+            return "null"
+
+        return json.encoder._make_iterencode(
+            markers,
+            self.default,
+            _encoder,
+            self.indent,
+            floatstr,
+            self.key_separator,
+            self.item_separator,
+            self.sort_keys,
+            self.skipkeys,
+            _one_shot,
+        )(o, 0)
 
 class EWAHCleaner(LoggingMixin):
     """Default data cleaner class for EWAH.
@@ -26,6 +104,7 @@ class EWAHCleaner(LoggingMixin):
         hash_columns: Optional[List[str]] = None,
         rename_columns: Optional[Dict[str, str]] = None,
         additional_callables: Optional[Union[List[Callable], Callable]] = None,
+        json_encoder: type = EWAHJSONEncoder,
     ):
         super().__init__()
 
@@ -58,6 +137,7 @@ class EWAHCleaner(LoggingMixin):
 
         self.cleaning_steps = cleaning_steps
         self.default_row = default_row or {}
+        self.json_encoder = json_encoder
 
         if default_row:
             # initialize with defaults
@@ -118,6 +198,16 @@ class EWAHCleaner(LoggingMixin):
                         # Some database systems don't handle this character well
                         # Thus, remove it
                         value = value.replace("\x00", "")
+                elif isinstance(value, (dict, OrderedDict, list)):
+                    # Logic copy-pasted from legacy - TODO: Refactor this!
+                    try:
+                        value = json.dumps(
+                            value, cls=self.json_encoder
+                        )
+                    except TypeError:
+                        # try dumping with bson utility function
+                        # Refactor this, PLEASE!
+                        value = dumps(value)
                 row[key] = value
 
                 # Set the fields_definition for the key
